+++ crates/mako/src/generate/optimize_chunk.rs
@@ -3,6 +3,7 @@ use std::string::String;
 
 use base64::engine::general_purpose;
 use base64::Engine;
+use hashlink::LinkedHashSet;
 use indexmap::{IndexMap, IndexSet};
 use regex::Regex;
 use tracing::debug;
@@ -40,9 +41,6 @@ impl Compiler {
                 .collect::<Vec<_>>();
 
             optimize_chunks_infos.sort_by_key(|o| -o.group_options.priority);
-
-            // stage: deasync
-            self.merge_minimal_async_chunks(&optimize_options);
 
             // stage: modules
             self.module_to_optimize_infos(&mut optimize_chunks_infos, None);
@@ -117,58 +115,6 @@ impl Compiler {
         }
     }
 
-    fn merge_minimal_async_chunks(&self, options: &CodeSplittingAdvancedOptions) {
-        let mut async_to_entry = vec![];
-        let chunk_graph = self.context.chunk_graph.read().unwrap();
-        let chunks = chunk_graph.get_all_chunks();
-
-        // find minimal async chunks to merge to entry chunk
-        // TODO: continue to merge deep-level async chunk
-        for chunk in chunks {
-            if chunk.chunk_type == ChunkType::Async && self.get_chunk_size(chunk) < options.min_size
-            {
-                let entry_ids = chunk_graph.entry_dependents_chunk(&chunk.id);
-
-                // merge if there is only one entry chunk
-                // TODO: don't merge if entry chunk size is greater than max_size
-                if entry_ids.len() == 1 {
-                    async_to_entry.push((
-                        chunk.id.clone(),
-                        entry_ids[0].clone(),
-                        chunk.modules.iter().cloned().collect::<Vec<_>>(),
-                    ));
-                }
-            }
-        }
-        drop(chunk_graph);
-
-        // update chunk_graph
-        let mut chunk_graph = self.context.chunk_graph.write().unwrap();
-        let mut merged_modules = vec![];
-
-        for (chunk_id, entry_chunk_id, chunk_modules) in async_to_entry.iter() {
-            let entry_chunk: &mut Chunk = chunk_graph.mut_chunk(entry_chunk_id).unwrap();
-
-            // merge modules to entry chunk
-            for m in chunk_modules {
-                entry_chunk.add_module(m.clone());
-                merged_modules.push(m);
-            }
-
-            // remove original async chunks
-            chunk_graph.merge_to_chunk(chunk_id, entry_chunk_id);
-        }
-
-        // remove merged modules from other async chunks
-        let mut chunks = chunk_graph.mut_chunks();
-
-        for chunk in chunks.iter_mut() {
-            if chunk.chunk_type == ChunkType::Async {
-                chunk.modules.retain(|m| !merged_modules.contains(&m));
-            }
-        }
-    }
-
     fn module_to_optimize_infos<'a>(
         &'a self,
         optimize_chunks_infos: &'a mut Vec<OptimizeChunksInfo>,
@@ -179,7 +125,7 @@ impl Compiler {
         let async_chunk_root_modules = chunks
             .iter()
             .filter_map(|chunk| match chunk.chunk_type {
-                ChunkType::Async => chunk.modules.last(),
+                ChunkType::Async => chunk.modules.iter().last(),
                 _ => None,
             })
             .collect::<Vec<_>>();
@@ -267,7 +213,7 @@ impl Compiler {
                         .module_to_chunks
                         .keys()
                         .cloned()
-                        .collect::<IndexSet<_>>(),
+                        .collect::<LinkedHashSet<_>>(),
                     id: ChunkId { id: "".to_string() },
                     chunk_type: ChunkType::Sync,
                     content: None,
@@ -492,7 +438,7 @@ impl Compiler {
                     .module_to_chunks
                     .keys()
                     .cloned()
-                    .collect::<IndexSet<_>>(),
+                    .collect::<LinkedHashSet<_>>(),
                 id: info_chunk_id.clone(),
                 chunk_type: info_chunk_type,
                 content: None,