+++ e2e/fixtures/sass.normal/src/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import styles from "./index.scss";
+
+export default () => {
+  return <div className={styles.title}>working with Sass</div>;
+};